+++ frontend/src/main.jsx
@@ -2,7 +2,7 @@ import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import "./index.css";
 import App from "./App.jsx";
-import Login from "./pages/Login.jsx";
+import Login from "./pages/Signup.jsx";
 import { BrowserRouter, Route, Routes } from "react-router";
 import Layout from "./components/Layout.jsx";
 import TestChart from "./components/charts/TestChart.jsx";
@@ -16,6 +16,7 @@ import "/node_modules/react-grid-layout/css/styles.css";
 import "/node_modules/react-resizable/css/styles.css";
 import Alarms from "./pages/Alarms"; 
 import Anamolies from "./pages/Anamolies.jsx";
+import SignIn from "./pages/SignIn.jsx";
 
 
 createRoot(document.getElementById("root")).render(
@@ -24,6 +25,7 @@ createRoot(document.getElementById("root")).render(
       <BrowserRouter>
         <Routes>
         <Route index element={<Login />} />
+        <Route path="/hello" element={<SignIn />}/>
           <Route path="/" element={<Layout />}>
             
             {/* <Route path="testchart" element={<TestChart />} /> */}
